+++ lending/src/sdk/models/shared/accountingaccounttransaction.ts
@@ -11,7 +11,7 @@ import { Expose, Type } from "class-transformer";
 /**
  * The status of the account transaction.
  */
-export enum AccountingAccountTransactionStatus {
+export enum Status {
     Unknown = "Unknown",
     Unreconciled = "Unreconciled",
     Reconciled = "Reconciled",
@@ -91,6 +91,13 @@ export class AccountingAccountTransaction extends SpeakeasyBase {
      * | **GBP**          | £20            | 1.277         | $25.54                     |
      * | **EUR**          | €20            | 1.134         | $22.68                     |
      * | **RUB**          | ₽20            | 0.015         | $0.30                      |
+     *
+     *
+     * ### Integration-specific details
+     *
+     * | Integration       | Scenario                                        | System behavior                                                                                                                                                      |
+     * |-------------------|-------------------------------------------------|----------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+     * | QuickBooks Online | Transaction currency differs from base currency | If currency rate value is left `null`, a rate of 1 will be used by QBO by default. To override this, include the required currency rate in the expense transaction.  |
      */
     @SpeakeasyMetadata()
     @Expose({ name: "currencyRate" })
@@ -163,7 +170,7 @@ export class AccountingAccountTransaction extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     @Expose({ name: "status" })
-    status?: AccountingAccountTransactionStatus;
+    status?: Status;
 
     /**
      * Total amount of the account transactions, inclusive of tax.