+++ lending/src/sdk/models/shared/accountinginvoice.ts
@@ -8,12 +8,12 @@ import { AccountingPaymentAllocation } from "./accountingpaymentallocation";
 import { DataType } from "./datatype";
 import { InvoiceLineItem } from "./invoicelineitem";
 import { InvoiceStatus } from "./invoicestatus";
-import { Items } from "./items";
 import { Metadata } from "./metadata";
 import { SupplementalData } from "./supplementaldata";
+import { WithholdingTaxItems } from "./withholdingtaxitems";
 import { Expose, Type } from "class-transformer";
 
-export class AccountingInvoiceSalesOrderReference extends SpeakeasyBase {
+export class SalesOrderReference extends SpeakeasyBase {
     /**
      * Available Data types
      */
@@ -127,6 +127,13 @@ export class AccountingInvoice extends SpeakeasyBase {
      * | **GBP**          | £20            | 1.277         | $25.54                     |
      * | **EUR**          | €20            | 1.134         | $22.68                     |
      * | **RUB**          | ₽20            | 0.015         | $0.30                      |
+     *
+     *
+     * ### Integration-specific details
+     *
+     * | Integration       | Scenario                                        | System behavior                                                                                                                                                      |
+     * |-------------------|-------------------------------------------------|----------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+     * | QuickBooks Online | Transaction currency differs from base currency | If currency rate value is left `null`, a rate of 1 will be used by QBO by default. To override this, include the required currency rate in the expense transaction.  |
      */
     @SpeakeasyMetadata()
     @Expose({ name: "currencyRate" })
@@ -274,10 +281,10 @@ export class AccountingInvoice extends SpeakeasyBase {
     /**
      * List of references to related Sales orders.
      */
-    @SpeakeasyMetadata({ elemType: AccountingInvoiceSalesOrderReference })
+    @SpeakeasyMetadata({ elemType: SalesOrderReference })
     @Expose({ name: "salesOrderRefs" })
-    @Type(() => AccountingInvoiceSalesOrderReference)
-    salesOrderRefs?: AccountingInvoiceSalesOrderReference[];
+    @Type(() => SalesOrderReference)
+    salesOrderRefs?: SalesOrderReference[];
 
     @SpeakeasyMetadata()
     @Expose({ name: "sourceModifiedDate" })
@@ -338,8 +345,8 @@ export class AccountingInvoice extends SpeakeasyBase {
     @Expose({ name: "totalTaxAmount" })
     totalTaxAmount: number;
 
-    @SpeakeasyMetadata({ elemType: Items })
+    @SpeakeasyMetadata({ elemType: WithholdingTaxItems })
     @Expose({ name: "withholdingTax" })
-    @Type(() => Items)
-    withholdingTax?: Items[];
+    @Type(() => WithholdingTaxItems)
+    withholdingTax?: WithholdingTaxItems[];
 }