+++ lending/src/sdk/models/shared/connection.ts
@@ -10,10 +10,12 @@ import { Expose, Type } from "class-transformer";
 /**
  * The type of platform of the connection.
  */
-export enum ConnectionSourceType {
+export enum SourceType {
     Accounting = "Accounting",
     Banking = "Banking",
+    BankFeed = "BankFeed",
     Commerce = "Commerce",
+    Expense = "Expense",
     Other = "Other",
     Unknown = "Unknown",
 }
@@ -147,7 +149,7 @@ export class Connection extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     @Expose({ name: "sourceType" })
-    sourceType: ConnectionSourceType;
+    sourceType: SourceType;
 
     /**
      * The current authorization status of the data connection.